requirements.txt
streamlit
pydantic
openai
PyPDF2
python-docx
newspaper3k
lxml
youtube-transcript-api
python-dotenv
